Move main platform to '(app)' folder
