+++ app/(app)/create/[[...paramsArr]]/_client.tsx
@@ -3,20 +3,20 @@ import * as Sentry from "@sentry/nextjs";
 import { ZodError } from "zod";
 import React, { useState, useEffect, Fragment, useRef } from "react";
 import { useForm } from "react-hook-form";
-import CustomTextareaAutosize from "../../../components/CustomTextareAutosize/CustomTextareaAutosize";
+import CustomTextareaAutosize from "@/components/CustomTextareAutosize/CustomTextareaAutosize";
 import { toast } from "sonner";
 import { Disclosure, Transition } from "@headlessui/react";
 import { ChevronUpIcon } from "@heroicons/react/20/solid";
-import type { SavePostInput } from "../../../schema/post";
-import { ConfirmPostSchema } from "../../../schema/post";
+import type { SavePostInput } from "@/schema/post";
+import { ConfirmPostSchema } from "@/schema/post";
 import { api } from "@/server/trpc/react";
-import { removeMarkdown } from "../../../utils/removeMarkdown";
-import { useDebounce } from "../../../hooks/useDebounce";
+import { removeMarkdown } from "@/utils/removeMarkdown";
+import { useDebounce } from "@/hooks/useDebounce";
 import Markdoc from "@markdoc/markdoc";
-import { useMarkdownHotkeys } from "../../../markdoc/editor/hotkeys/hotkeys.markdoc";
-import { useMarkdownShortcuts } from "../../../markdoc/editor/shortcuts/shortcuts.markdoc";
-import { markdocComponents } from "../../../markdoc/components";
-import { config } from "../../../markdoc/config";
+import { useMarkdownHotkeys } from "@/markdoc/editor/hotkeys/hotkeys.markdoc";
+import { useMarkdownShortcuts } from "@/markdoc/editor/shortcuts/shortcuts.markdoc";
+import { markdocComponents } from "@/markdoc/components";
+import { config } from "@/markdoc/config";
 import { useParams, useRouter } from "next/navigation";
 import { usePrompt } from "@/components/PromptService";
 import { Switch } from "@/components/Switch/Switch";
