+++ app/(app)/hub/[community]/edit/page.tsx
@@ -1,9 +1,13 @@
-import { EventForm } from "@/components/EventForm/EventForm";
-import prisma from "../../../../../server/db/client";
+import { CommunityForm } from "@/components/CommunityForm/CommunityForm";
+import prisma from "@/server/db/client";
 import { getServerAuthSession } from "@/server/auth";
-import { redirect, notFound } from "next/navigation";
+import { notFound, redirect } from "next/navigation";
 
-async function CreateEventPage({ params }: { params: { community: string } }) {
+async function EditCommunityPage({
+  params,
+}: {
+  params: { community: string };
+}) {
   const session = await getServerAuthSession();
 
   if (!session) {
@@ -20,10 +24,17 @@ async function CreateEventPage({ params }: { params: { community: string } }) {
     },
     include: {
       members: {
-        select: {
-          id: true,
-          isEventOrganiser: true,
-          userId: true,
+        include: {
+          user: true,
+        },
+      },
+      events: {
+        include: {
+          RSVP: {
+            select: {
+              id: true,
+            },
+          },
         },
       },
     },
@@ -41,18 +52,20 @@ async function CreateEventPage({ params }: { params: { community: string } }) {
     redirect("/forbidden");
   }
 
+  const { id, name, city, country, coverImage, description, excerpt } =
+    community;
   return (
     <div className="mx-auto flex w-full max-w-2xl flex-grow flex-col justify-center px-4 sm:px-6 lg:col-span-9">
       <div className="bg-neutral-900 text-neutral-700 shadow-xl">
-        <EventForm
+        <CommunityForm
           defaultValues={{
-            address: "",
-            description: "",
-            name: "",
-            capacity: 50,
-            eventDate: new Date(),
-            communityId: community.id,
-            coverImage: "",
+            id,
+            name,
+            city,
+            country,
+            coverImage: coverImage ? coverImage : "",
+            description,
+            excerpt,
           }}
         />
       </div>
@@ -60,4 +73,4 @@ async function CreateEventPage({ params }: { params: { community: string } }) {
   );
 }
 
-export default CreateEventPage;
+export default EditCommunityPage;
