+++ app/(app)/hub/[community]/events/[event]/_EventPage.tsx
