+++ app/(app)/my-posts/_client.tsx
