+++ app/(app)/my-posts/page.tsx
