+++ app/(app)/settings/page.tsx
@@ -2,7 +2,7 @@ import { redirect } from "next/navigation";
 import Content from "./_client";
 import { getServerAuthSession } from "@/server/auth";
 import { customAlphabet } from "nanoid";
-import prisma from "../../server/db/client";
+import prisma from "@/server/db/client";
 
 export const metadata = {
   title: "Settings - Update your profile",
